+++ packages/contracts/src/2.0.0/protocol/Exchange/MixinTransactions.sol
@@ -28,22 +28,12 @@ contract MixinTransactions is
     MSignatureValidator,
     MTransactions
 {
-
     // Mapping of transaction hash => executed
     // This prevents transactions from being executed more than once.
     mapping (bytes32 => bool) public transactions;
 
     // Address of current transaction signer
     address public currentContextAddress;
-
-    // Hash for the EIP712 ZeroEx Transaction Schema
-    bytes32 constant internal EIP712_ZEROEX_TRANSACTION_SCHEMA_HASH = keccak256(abi.encodePacked(
-        "ZeroExTransaction(",
-        "uint256 salt,",
-        "address signerAddress,",
-        "bytes data",
-        ")"
-    ));
 
     /// @dev Executes an exchange method call in the context of signer.
     /// @param salt Arbitrary number to ensure uniqueness of transaction hash.