+++ packages/contracts/src/2.0.0/protocol/Exchange/mixins/MWrapperFunctions.sol
@@ -24,8 +24,9 @@ import "../libs/LibFillResults.sol";
 import "../interfaces/IWrapperFunctions.sol";
 
 
-contract MWrapperFunctions {
-
+contract MWrapperFunctions is 
+    IWrapperFunctions
+{
     /// @dev Fills the input order. Reverts if exact takerAssetFillAmount not filled.
     /// @param order LibOrder.Order struct containing order specifications.
     /// @param takerAssetFillAmount Desired amount of takerAsset to sell.