+++ packages/contracts/src/2.0.0/test/DummyERC20Token/DummyNoReturnERC20Token.sol
@@ -25,7 +25,6 @@ import "./DummyERC20Token.sol";
 contract DummyNoReturnERC20Token is
     DummyERC20Token
 {
-
     constructor (
         string _name,
         string _symbol,