+++ packages/contracts/src/2.0.0/test/TestAssetProxyOwner/TestAssetProxyOwner.sol
@@ -25,7 +25,6 @@ import "../../protocol/AssetProxyOwner/AssetProxyOwner.sol";
 contract TestAssetProxyOwner is
     AssetProxyOwner
 {
-
     constructor (
         address[] memory _owners,
         address[] memory _assetProxyContracts,