+++ packages/contracts/src/2.0.0/test/TestLibs/TestLibs.sol
@@ -31,7 +31,6 @@ contract TestLibs is
     LibFillResults,
     LibAbiEncoder
 {
-
     function publicAbiEncodeFillOrder(
         Order memory order,
         uint256 takerAssetFillAmount,