+++ packages/migrations/src/utils/token_info.ts
@@ -2,6 +2,20 @@ import { BigNumber, NULL_BYTES } from '@0xproject/utils';
 
 import { ERC20Token, ERC721Token } from '../types';
 
+export const etherTokenByNetwork: { [networkId: number]: { address: string } } = {
+    3: {
+        address: '0xc778417e063141139fce010982780140aa0cd5ab',
+    },
+    4: {
+        address: '0xc778417e063141139fce010982780140aa0cd5ab',
+    },
+    42: {
+        address: '0xd0a1e359811322d97991e03f863a0c30c2cf029c',
+    },
+    50: {
+        address: '',
+    },
+};
 export const erc20TokenInfo: ERC20Token[] = [
     {
         name: 'Augur Reputation Token',
@@ -42,7 +56,7 @@ export const erc20TokenInfo: ERC20Token[] = [
 
 export const erc721TokenInfo: ERC721Token[] = [
     {
-        name: 'CryptoKitties',
-        symbol: 'CK',
+        name: '0xen ERC721',
+        symbol: '0xen',
     },
 ];