+++ packages/order-watcher/CHANGELOG.md
@@ -5,6 +5,10 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v1.0.1 - _September 5, 2018_
+
+    * Dependencies updated
+
 ## v1.0.1-rc.5 - _August 27, 2018_
 
     * Fix missing `BlockParamLiteral` type import issue